Bug Fixes:
1.Crash in Home Screen Banner Click:  FIXED
When clicking on a banner on the home screen, it produces an ActivityNotFoundException when trying to open the URL. The banner link from the API lacks the protocol part (https), which is the reason for the exception. I have handled the exception from the code side, and we need to change the URL format from "www.welkinwitstech.com/" to "https://www.welkinwitstech.com/".

2.Images Are Not Showing for News and Evening Sections: FIXED
The image URLs loaded from the backend are not valid:

https://app.witsclass.com/console/DEVOPS/micros/public_html/images/img/1714202102_1743unnamed.jpg
https://app.witsclass.com/console/DEVOPS/micros/public_html/images/img/man.jpg

3.Change the Click Action for Profile in the Footer. :FIXED

4.Remove Subscription Feature: Hide the subscription feature from Examination, Recorded Class, and Library. : FIXED